Implementation of lookupVar, updateVar, and changes in the environment.
The environemnt vars are now a map from Var to [String], as to allow for
notation such as
  let $xs = [ls, ls -al, echo "hello"] in { echo i : i <- $xs }

Adding some debug prints.

Adding a go on implementing let expressions, seems we are missing some
functionality to allow this. We need to handle the stdout better. I saw that it
is possible to pass environments to CreateProcess, as well as, working
directory, std_in, std_out, std_err etc, seems promising.

Adding updateVar, which updates a var in the environment.

Adding lookupVar, which looks up a var in the environment. Errors are undefined
as for now.

Change concatText so that the get method looks up a var in the environment if
the Text is a TVar, and simply returns the [String] if it is a TLit. Not sure if
the concat is correct, but it seems to work.

I'm not quite sure that the value of a var in the env ought to be a [String],
instead of String. But we will figure that out.