Implement redirection of input/output
The I/O can now be redirected to files. The current working dir is a bit
strange as the shell commands are executed with one path while the
haskell ones use another. Should be fixed when implementing cd.

The grammar is a bit restrictive, wc < testfile | ls doesn't work
without parantheses, it must be (wc < testfile) | ls.