Implement list comprehensions on the form { Exp : Var <- Exp }
The variable is assigned each of the values in the output of the second
expression, so {echo $x : $x <- ls} will print each filename on its own
line.

Maybe we should change the behaviour of the other list comprehension
(with '[]') to not evaluate its arguments but instead just treat them as
text?

For example:
{ echo $x : $x <- [a, b, hej]}
will evaluate to:
a
b
hej

Which may be convenient in some cases. This would also make sense if we
wanted lists such as [1..10].