Use only Lit in the grammar. Make let's work
We know use only Lit in the grammar, but have added a "string"
expression, that will evaluate only to itself. This way, we can have for
example:

let x = "test" in cat $x/hello

But also:

let x = ls | wc in echo $x
or:
let x = "ls" in $x
Work as they should.