Change the behaviour of (()) to be more like bash.
If a wrap is the first part of an expression (the binary) it is
evaluated to a string as it is and then reinterpreted along with the
rest of the wraps as its arguments. The rest of the wraps are evaluated
to strings as usual, but if they are (())'s, they are wrapped with "" to
make them literals. This way, we can run: let x = ls - l echo (($x)) $x

Which will echo both the output of ls -l and the string ls -l itself.

Updated the testscripts to reflect this change